+++ src/main/java/com/stocks/aggregator/configuration/RestConfig.java
@@ -5,7 +5,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
 
 @Configuration
-public class AppConfig {
+public class RestConfig {
 
     @Bean
     public RestTemplate restTemplate() {
