+++ src/main/java/com/stocks/aggregator/etoro/AccountActivityService.java
@@ -1,8 +1,8 @@
-package com.stocks.aggregator.service;
+package com.stocks.aggregator.etoro;
 
 
-import com.stocks.aggregator.db.repository.AccountActivityRepository;
-import com.stocks.aggregator.model.etoro.AccountActivity;
+import com.stocks.aggregator.etoro.repo.AccountActivityRepository;
+import com.stocks.aggregator.etoro.model.AccountActivity;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
