+++ src/main/java/com/stocks/aggregator/etoro/DayTradeStatusService.java
@@ -1,10 +1,10 @@
-package com.stocks.aggregator.service;
+package com.stocks.aggregator.etoro;
 
-import com.stocks.aggregator.db.repository.ClosedTradePositionRepository;
-import com.stocks.aggregator.db.repository.DayTradeStatusRepository;
-import com.stocks.aggregator.model.DayTradeStatus;
-import com.stocks.aggregator.model.etoro.AccountActivity;
-import com.stocks.aggregator.model.etoro.ClosedTradePosition;
+import com.stocks.aggregator.etoro.repo.ClosedTradePositionRepository;
+import com.stocks.aggregator.etoro.repo.DayTradeStatusRepository;
+import com.stocks.aggregator.etoro.model.DayTradeStatus;
+import com.stocks.aggregator.etoro.model.AccountActivity;
+import com.stocks.aggregator.etoro.model.ClosedTradePosition;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -15,7 +15,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicReference;
 
-import static com.stocks.aggregator.service.ClosedPositionService.*;
+import static com.stocks.aggregator.etoro.ClosedPositionService.*;
 
 @Service
 @AllArgsConstructor
