+++ src/main/java/com/stocks/aggregator/etoro/repo/MonthTradeStatusRepository.java
@@ -1,7 +1,7 @@
-package com.stocks.aggregator.db.repository;
+package com.stocks.aggregator.etoro.repo;
 
 
-import com.stocks.aggregator.model.MonthTradeStatus;
+import com.stocks.aggregator.etoro.model.MonthTradeStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MonthTradeStatusRepository extends JpaRepository<MonthTradeStatus, Long> {
