+++ src/main/java/com/stocks/aggregator/global/User.java
@@ -1,4 +1,4 @@
-package com.stocks.aggregator.model;
+package com.stocks.aggregator.global;
 
 import jakarta.persistence.*;
 import lombok.Data;
