+++ src/main/java/com/stocks/aggregator/global/repo/UserRepository.java
@@ -1,7 +1,7 @@
-package com.stocks.aggregator.db.repository;
+package com.stocks.aggregator.global.repo;
 
 
-import com.stocks.aggregator.model.User;
+import com.stocks.aggregator.global.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
