+++ src/main/java/com/stocks/aggregator/revolut/RevolutStatementService.java
@@ -1,6 +1,8 @@
 package com.stocks.aggregator.revolut;
 
-import com.stocks.aggregator.db.repository.RevolutStatementRepository;
+import com.stocks.aggregator.revolut.repo.RevolutStatementRepository;
+import com.stocks.aggregator.revolut.model.RevolutStatement;
+import com.stocks.aggregator.revolut.model.RevolutTransactionType;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
